+++ attributes/composer.rb
@@ -1,6 +1,6 @@
 #
 # Attributes controlling the global composer installation
-
-default['composer']['global_cache']['path'] = '/var/composer/cache'
-default['composer']['global_cache']['user']  = 'root'
-default['composer']['global_cache']['group'] = 'root'
+default['composer']['binary_path'] = '/usr/local/bin/composer'
+default['composer']['global_cache_dir'] = '/var/composer/cache'
+default['composer']['optimize-autoloader?'] = node.not_environment?(:localdev)
+default['composer']['apcu-autoloader?'] = node.not_environment?(:localdev)