+++ metadata.rb
@@ -6,12 +6,11 @@ description 'Installs PHP with standard config and helpers for inGenerator proje
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 issues_url 'https://github.com/ingenerator/chef-ingenerator-php/issues'
 source_url 'https://github.com/ingenerator/chef-ingenerator-php'
-version '1.0.0-pre'
+version '1.0.0'
 
 %w(ubuntu).each do |os|
   supports os
 end
 
-depends 'composer'
 depends 'ingenerator-helpers', '~> 1.0'
 depends 'php', '~> 2.2'