+++ recipes/composer.rb
@@ -3,32 +3,11 @@
 # in /var/composer/cache. Note that your composer.json needs to be set to use
 # this path separately.
 #
-# Author::  Andrew Coulton (<andrew@ingenerator.com>)
-# Cookbook Name:: ingenerator-php
-# Recipe:: install_php
-#
-# Copyright 2012-13, inGenerator Ltd
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
 
-composer node['composer']['install_path'] do
+# Presence of this attribute suggests the project is still using zircote/chef-composer
+# which conflicts with our new resources.
+raise_if_legacy_attributes('composer.install_path', 'composer.global_cache')
+
+composer_binary node['composer']['binary_path'] do
   action [:install, :update]
-end
-
-directory node['composer']['global_cache']['path'] do
-  recursive true
-  user      node['composer']['global_cache']['user']
-  group     node['composer']['global_cache']['group']
-  mode      0777
 end