+++ test/cookbooks/test_helpers/recipes/test_composer_binary.rb
@@ -1,0 +1,3 @@
+composer_binary node['test']['path'] do
+  action node['test']['action']
+end