+++ test/cookbooks/test_helpers/recipes/test_composer_dependencies.rb
@@ -1,0 +1,3 @@
+composer_dependencies node['test']['project_dir'] do
+  run_as node['test']['run_as']
+end